requirements.txt
pvporcupine==2.1.4
PyAudio==0.2.13
baidu-aip==4.16.10
SpeechRecognition==3.10.0
playsound==1.3.0
pyttsx3==2.90
azure-cognitiveservices-speech==1.27.0
openai==0.27.4
